deprecate duplicate methods, point to alts, and split out existing tests
