+++ openlibrary/catalog/merge/merge.py
@@ -1,11 +1,10 @@
-from __future__ import print_function
 import re
-from names import match_name
-from normalize import normalize
+import warnings
 
-import six
+from openlibrary.catalog.merge.names import match_name
+from openlibrary.catalog.merge.normalize import normalize
 
-re_year = re.compile('(\d{4})$')
+
 re_amazon_title_paren = re.compile('^(.*) \([^)]+?\)$')
 re_and_of_space = re.compile(' and | of | ')
 
@@ -13,34 +12,6 @@ isbn_match = 85
 
 def set_isbn_match(score):
     isbn_match = score
-
-def amazon_year(date):
-    m = re_year.search(date)
-    try:
-        assert m
-        year = m.group(1)
-    except:
-        print(date)
-        raise
-    return year
-
-def build_amazon(edition, author):
-    amazon = merge.build_titles(full_title(edition))
-
-    amazon['isbn'] = editon['isbn_10']
-    if 'publish_date' in edition:
-        amazon['publish_date'] = merge.amazon_year(edition['publish_date'])
-    if author:
-        amazon['authors'] = author
-    else:
-        amazon['authors'] = []
-    if 'number_of_pages' in edition:
-        amazon['number_of_pages'] = edition['number_of_pages']
-    if 'publishers' in edition:
-        assert len(edition['publishers']) == 1
-        amazon['publisher'] = edition['publishers'][0]
-
-    return amazon
 
 
 def build_titles(title):
@@ -94,7 +65,10 @@ def compare_date(e1, e2):
     except ValueError as TypeError:
         return ('date', 'mismatch', -250)
 
+
 def compare_isbn10(e1, e2):
+    warnings.warn('Deprecated, use openlibrary.catalog.merge.merge_marc.compare_isbn10() instead.', DeprecationWarning)
+
     if len(e1['isbn']) == 0 or len(e2['isbn']) == 0:
         return ('ISBN', 'missing', 0)
     for i in e1['isbn']:
@@ -105,6 +79,8 @@ def compare_isbn10(e1, e2):
     return ('ISBN', 'mismatch', -225)
 
 def level1_merge(e1, e2):
+    warnings.warn('Deprecated, use openlibrary.catalog.merge.merge_marc.level1_merge() instead.', DeprecationWarning)
+
     score = []
     if e1['short_title'] == e2['short_title']:
         score.append(('short-title', 'match', 450))
@@ -116,6 +92,8 @@ def level1_merge(e1, e2):
     return score
 
 def compare_authors(amazon, marc):
+    warnings.warn('Deprecated, use openlibrary.catalog.merge.merge_marc.compare_authors() instead.', DeprecationWarning)
+
     if len(amazon['authors']) == 0 and 'authors' not in marc:
         return ('main', 'no authors', 75)
     if len(amazon['authors']) == 0:
@@ -182,7 +160,7 @@ def compare_title(amazon, marc):
         for a in amazon['titles']:
             for m in marc['titles']:
                 if substr_match(a.lower(), m.lower()):
-                    return ('full-title', 'containted within other title', 350)
+                    return ('full-title', 'contained within other title', 350)
 
     max_score = 0
     for a in amazon['titles']:
@@ -248,7 +226,8 @@ def compare_publisher(amazon, marc):
         return ('publisher', 'either missing', 0)
 
 def level2_merge(amazon, marc):
-#    print 'MARC', marc
+    warnings.warn('Deprecated, use openlibrary.catalog.merge.merge_marc.level2_merge() instead.', DeprecationWarning)
+
     score = []
     score.append(compare_date(amazon, marc))
     score.append(compare_isbn10(amazon, marc))
@@ -259,7 +238,6 @@ def level2_merge(amazon, marc):
 
     score.append(compare_publisher(amazon, marc))
     score.append(compare_authors(amazon, marc))
-
     return score
 
 def full_title(edition):
@@ -268,42 +246,9 @@ def full_title(edition):
         title += ' ' + edition['subtitle']
     return title
 
-def test_full_title():
-    assert full_title({ 'title': "Hamlet"}) == "Hamlet"
-    edition = {
-        'title': 'Flatland',
-        'subtitle': 'A Romance of Many Dimensions',
-    }
-    assert full_title(edition) == "Flatland A Romance of Many Dimensions"
-
-def test_merge_titles():
-    marc = {
-        'title_with_subtitles': 'Spytime : the undoing of James Jesus Angleton : a novel',
-        'title': 'Spytime',
-        'full_title': 'Spytime : the undoing of James Jesus Angleton : a novel',
-    }
-    amazon = {
-        'subtitle': 'The Undoing oF James Jesus Angleton',
-        'title': 'Spytime',
-    }
-
-    amazon = build_titles(six.text_type(full_title(amazon)))
-    marc = build_titles(marc['title_with_subtitles'])
-    assert amazon['short_title'] == marc['short_title']
-    assert compare_title(amazon, marc) == ('full-title', 'containted within other title', 350)
-
-def test_merge_titles2():
-    amazon = {'title': u'Sea Birds Britain Ireland'}
-    marc = {
-        'title_with_subtitles': u'seabirds of Britain and Ireland',
-        'title': u'seabirds of Britain and Ireland',
-        'full_title': u'The seabirds of Britain and Ireland',
-    }
-    amazon = build_titles(six.text_type(full_title(amazon)))
-    marc = build_titles(marc['title_with_subtitles'])
-    assert compare_title(amazon, marc) == ('full-title', 'exact match', 600)
-
 def attempt_merge(amazon, marc, threshold):
+    warnings.warn('Deprecated, use openlibrary.catalog.merge.merge_marc.attempt_merge() instead.', DeprecationWarning)
+
     l1 = level1_merge(amazon, marc)
     total = sum(i[2] for i in l1)
     if total >= threshold:
@@ -311,66 +256,3 @@ def attempt_merge(amazon, marc, threshold):
     l2 = level2_merge(amazon, marc)
     total = sum(i[2] for i in l2)
     return total >= threshold
-
-def test_merge():
-    amazon = {'publisher': u'Collins', 'isbn_10': ['0002167360'], 'number_of_pages': 120, 'short_title': u'souvenirs', 'normalized_title': u'souvenirs', 'full_title': u'Souvenirs', 'titles': [u'Souvenirs', u'souvenirs'], 'publish_date': u'1975', 'authors': [(u'David Hamilton', u'Photographer')]}
-    marc = {'publisher': [u'Collins'], 'isbn_10': [u'0002167360'], 'short_title': u'souvenirs', 'normalized_title': u'souvenirs', 'full_title': u'Souvenirs', 'titles': [u'Souvenirs', u'souvenirs'], 'publish_date': '1978', 'authors': [{'birth_date': u'1933', 'db_name': u'Hamilton, David 1933-', 'entity_type': 'person', 'name': u'Hamilton, David', 'personal_name': u'Hamilton, David'}], 'source_record_loc': 'marc_records_scriblio_net/part11.dat:155728070:617', 'number_of_pages': 120}
-
-    threshold = 735
-    assert attempt_merge(amazon, marc, threshold)
-
-def test_merge2():
-    amazon = {'publisher': u'Collins', 'isbn_10': ['0002167530'], 'number_of_pages': 287, 'short_title': u'sea birds britain ireland', 'normalized_title': u'sea birds britain ireland', 'full_title': u'Sea Birds Britain Ireland', 'titles': [u'Sea Birds Britain Ireland', u'sea birds britain ireland'], 'publish_date': u'1975', 'authors': [(u'Stanley Cramp', u'Author')]}
-    marc = {'publisher': [u'Collins'], 'isbn_10': [u'0002167530'], 'short_title': u'seabirds of britain and i', 'normalized_title': u'seabirds of britain and ireland', 'full_title': u'seabirds of Britain and Ireland', 'titles': [u'seabirds of Britain and Ireland', u'seabirds of britain and ireland'], 'publish_date': '1974', 'authors': [{'db_name': u'Cramp, Stanley.', 'entity_type': 'person', 'name': u'Cramp, Stanley.', 'personal_name': u'Cramp, Stanley.'}], 'source_record_loc': 'marc_records_scriblio_net/part08.dat:61449973:855'}
-
-    threshold = 735
-    #assert attempt_merge(amazon, marc, threshold)
-
-def test_merge3():
-    amazon = {'publisher': u'Intl Specialized Book Service Inc', 'isbn_10': ['0002169770'], 'number_of_pages': 207, 'short_title': u'women of the north', 'normalized_title': u'women of the north', 'full_title': u'Women of the North', 'titles': [u'Women of the North', u'women of the north'], 'publish_date': u'1985', 'authors': [(u'Jane Wordsworth', u'Author')]}
-    marc = {'publisher': [u'Collins', u'Exclusive distributor ISBS'], 'isbn_10': [u'0002169770'], 'short_title': u'women of the north', 'normalized_title': u'women of the north', 'full_title': u'Women of the North', 'titles': [u'Women of the North', u'women of the north'], 'publish_date': '1981', 'number_of_pages': 207, 'authors': [{'db_name': u'Wordsworth, Jane.', 'entity_type': 'person', 'name': u'Wordsworth, Jane.', 'personal_name': u'Wordsworth, Jane.'}], 'source_record_loc': 'marc_records_scriblio_net/part17.dat:110989084:798'}
-
-    threshold = 735
-#    assert attempt_merge(amazon, marc, threshold)
-
-def test_merge4():
-    amazon = {'publisher': u'HarperCollins Publishers Ltd', 'isbn_10': ['0002173433'], 'number_of_pages': 128, 'short_title': u'd day to victory', 'normalized_title': u'd day to victory', 'full_title': u'D-Day to Victory', 'titles': [u'D-Day to Victory', u'd day to victory'], 'publish_date': u'1984', 'authors': [(u'Wynfod Vaughan-Thomas', u'Editor, Introduction')]}
-    marc = {'publisher': [u'Collins'], 'isbn_10': [u'0002173433'], 'short_title': u'great front pages  d day ', 'normalized_title': u'great front pages  d day to victory 1944 1945', 'full_title': u'Great front pages : D-Day to victory 1944-1945', 'titles': [u'Great front pages : D-Day to victory 1944-1945', u'great front pages  dday to victory 1944 1945'], 'publish_date': '1984', 'number_of_pages': 128, 'by_statement': 'introduced by Wynford Vaughan-Thomas.', 'source_record_loc': 'marc_records_scriblio_net/part17.dat:102360356:983'}
-
-    threshold = 735
-    assert attempt_merge(amazon, marc, threshold)
-
-def test_merge5():
-    amazon = {'publisher': u'HarperCollins Publishers (Australia) Pty Ltd', 'isbn_10': ['0002174049'], 'number_of_pages': 120, 'short_title': u'netherlandish and german ', 'normalized_title': u'netherlandish and german paintings national gallery schools of painting', 'full_title': u'Netherlandish and German Paintings (National Gallery Schools of Painting)', 'titles': [u'Netherlandish and German Paintings (National Gallery Schools of Painting)', u'netherlandish and german paintings national gallery schools of painting', u'Netherlandish and German Paintings', u'netherlandish and german paintings'], 'publish_date': u'1985', 'authors': [(u'Alistair Smith', u'Author')]}
-    marc = {'publisher': [u'National Gallery in association with W. Collins'], 'isbn_10': [u'0002174049'], 'short_title': u'early netherlandish and g', 'normalized_title': u'early netherlandish and german paintings', 'full_title': u'Early Netherlandish and German paintings', 'titles': [u'Early Netherlandish and German paintings', u'early netherlandish and german paintings'], 'publish_date': '1985', 'authors': [{'db_name': u'National Gallery (Great Britain)', 'name': u'National Gallery (Great Britain)', 'entity_type': 'org'}], 'number_of_pages': 116, 'by_statement': 'Alistair Smith.', 'source_record_loc': 'marc_records_scriblio_net/part17.dat:170029527:1210'}
-    threshold = 735
-    assert attempt_merge(amazon, marc, threshold)
-
-def test_compare_authors():
-    amazon = {'authors': [(u'Alistair Smith', u'Author')]}
-    marc = {'authors': [{'db_name': u'National Gallery (Great Britain)', 'name': u'National Gallery (Great Britain)', 'entity_type': 'org'}], 'by_statement': 'Alistair Smith.'}
-    assert compare_authors(amazon, marc) == ('main', 'exact match', 125)
-
-def test_merge6():
-    amazon = {'publisher': u'Fount', 'isbn_10': ['0002176157'], 'number_of_pages': 224, 'short_title': u'basil hume', 'normalized_title': u'basil hume', 'full_title': u'Basil Hume', 'titles': [u'Basil Hume', u'basil hume'], 'publish_date': u'1986', 'authors': [(u'Tony Castle', u'Editor')]}
-    marc = {'publisher': [u'Collins'], 'isbn_10': [u'0002176157'], 'short_title': u'basil hume  a portrait', 'normalized_title': u'basil hume  a portrait', 'full_title': u'Basil Hume : a portrait', 'titles': [u'Basil Hume : a portrait', u'basil hume  a portrait'], 'number_of_pages': 158, 'publish_date': '1986', 'by_statement': 'edited by Tony Castle.', 'source_record_loc': 'marc_records_scriblio_net/part19.dat:39883132:951'}
-    threshold = 735
-    assert attempt_merge(amazon, marc, threshold)
-
-def test_merge7():
-    amazon = {'publisher': u'HarperCollins Publishers Ltd', 'isbn_10': ['0002176319'], 'number_of_pages': 256, 'short_title': u'pucklers progress', 'normalized_title': u'pucklers progress', 'full_title': u"Puckler's Progress", 'titles': [u"Puckler's Progress", u'pucklers progress'], 'publish_date': u'1987', 'authors': [(u'Flora Brennan', u'Editor')]}
-    marc = {'publisher': [u'Collins'], 'isbn_10': [u'0002176319'], 'short_title': u'pucklers progress  the ad', 'normalized_title': u'pucklers progress  the adventures of prince puckler muskau in england wales and ireland as told in letters to his former wife 1826 9', 'full_title': u"Puckler's progress : the adventures of Prince Pu\u0308ckler-Muskau in England, Wales, and Ireland as told in letters to his former wife, 1826-9", 'titles': [u"Puckler's progress : the adventures of Prince Pu\u0308ckler-Muskau in England, Wales, and Ireland as told in letters to his former wife, 1826-9", u'pucklers progress  the adventures of prince puckler muskau in england wales and ireland as told in letters to his former wife 1826 9'], 'publish_date': '1987', 'authors': [{'name': u'Pu\u0308ckler-Muskau, Hermann Furst von', 'title': u'Furst von', 'death_date': u'1871.', 'db_name': u'Pu\u0308ckler-Muskau, Hermann Furst von 1785-1871.', 'birth_date': u'1785', 'personal_name': u'Pu\u0308ckler-Muskau, Hermann', 'entity_type': 'person'}], 'number_of_pages': 254, 'by_statement': 'translated by Flora Brennan.', 'source_record_loc': 'marc_records_scriblio_net/part19.dat:148554594:1050'}
-    threshold = 735
-    assert attempt_merge(amazon, marc, threshold)
-
-def test_compare_publisher():
-    amazon = { 'publisher': 'foo' }
-    amazon2 = { 'publisher': 'bar' }
-    marc = { 'publishers': ['foo'] }
-    marc2 = { 'publishers': ['foo', 'bar'] }
-    assert compare_publisher({}, {}) == ('publisher', 'either missing', 0)
-    assert compare_publisher(amazon, {}) == ('publisher', 'either missing', 0)
-    assert compare_publisher({}, marc) == ('publisher', 'either missing', 0)
-    assert compare_publisher(amazon, marc) == ('publisher', 'match', 100)
-    assert compare_publisher(amazon2, marc) == ('publisher', 'mismatch', -25)
-    assert compare_publisher(amazon2, marc2) == ('publisher', 'match', 100)
